--- ruzhouxie-github.h
@@ -1,8 +1,0 @@
-﻿// ruzhouxie-github.h: 标准系统包含文件的包含文件
-// 或项目特定的包含文件。
-
-#pragma once
-
-#include <iostream>
-
-// TODO: 在此处引用程序需要的其他标头。
